Restructure media scripts.
